rename redux slices to be more consistent with other files
