+++ packages/client/src/redux/store.ts
@@ -1,8 +1,8 @@
 import { configureStore } from "@reduxjs/toolkit";
 import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
-import postReducer from "./slices/postSlice";
-import userReducer from "./slices/userSlice";
-import locationReducer from "./slices/locationSlice";
+import postReducer from "./slices/post-slice";
+import userReducer from "./slices/user-slice";
+import locationReducer from "./slices/location-slice";
 
 // TODO write instructions for setting up db.json
